ram vectors/irq: add basic support in linkerscript and start.s for ram vectors
RAM Vectors are address of irq vectors which are hold in ram.
The platform itself only support irq vectors in rom, but these aren't
good to changed when the system runs because we had to erase a block to
set irq.
Because of this there will be proxy irq handlers which do:
- check if ramvector addr == 0xffff -> rte (end irq handler)
- jump to the addr

Of corse this adds a higher latency to irq but the application programs
can be loaded afterwards and doesn't need to modified when used as
stand-alone application, because only it's the same method like the
platform irq handlers.